Implementing Bricks REPL (work in progress)
Moved to promise-based approach because blocking operations are not
possible in Node.js without native support for blocking operations
(there is a blocking file system I/O but there are no blocking sockets
I/O, timers etc.). The callbacks do not get processed by the JS Event
Loop until the currently executing function (that waits for a certain
condition via sleep) finishes and unwinds the stack. But if it
finishes, the blocking context is lost. For now, there is still much to
be re-implemented.